feature-iot-monitor/backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pydantic-settings==2.7.1
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.1
celery==5.4.0
redis==5.2.1
httpx==0.28.1
requests==2.32.0
psycopg2-binary==2.9.10
tenacity
pytest
pytest-mock